backend/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-community
langchain-groq
chromadb
pypdf
docx2txt
tiktoken
sentence-transformers
